+++ tests/testthat/test-distinctiveness_range.R
@@ -1,4 +1,4 @@
-context("Test Range-based Distinctiveness")
+# Test Range-based Distinctiveness
 
 # Initial data -----------------------------------------------------------------
 # Empty Matrix
@@ -119,8 +119,8 @@ good_di_0.9_ab = matrix(c(1/30, 7/90, NA,   NA,
 
 ## Sparse Matrices -------------------------------------------------------------
 library(Matrix)
-sparse_mat = as(valid_mat, "sparseMatrix")
-sparse_ab_mat = as(abund_mat, "sparseMatrix")
+sparse_mat = as(valid_mat, "dgCMatrix")
+sparse_ab_mat = as(abund_mat, "dgCMatrix")
 
 # Tests for Distinctiveness ----------------------------------------------------
 
@@ -130,14 +130,6 @@ test_that("Bad input generates error", {
 
   expect_error(distinctiveness_range(valid_mat, dist_mat, 0.1, relative = "A"),
                "'relative' argument should be either TRUE or FALSE")
-
-  expect_warning(
-    distinctiveness_range(valid_mat * 4, dist_mat, 0.1),
-    paste0("Provided object may not contain relative abundances nor ",
-           "presence-absence\n",
-           "Have a look at the make_relative() function if it is the case"),
-    fixed = TRUE
-  )
 })
 
 test_that("Correct Di computation without abundance",{
@@ -148,23 +140,23 @@ test_that("Correct Di computation without abundance",{
 
   # Check computation equal
   # When T < 1/9
-  expect_equivalent(distinctiveness_range(valid_mat, dist_mat, 0.1),
-                    good_di_0.1)
+  expect_equal(distinctiveness_range(valid_mat, dist_mat, 0.1),
+               good_di_0.1)
 
   # 1/9 ≤ T < 4/9
-  expect_equivalent(distinctiveness_range(valid_mat, dist_mat, 0.4),
-                    good_di_0.4)
+  expect_equal(distinctiveness_range(valid_mat, dist_mat, 0.4),
+               good_di_0.4)
 
   # 4/9 ≤ T < 8/9
-  expect_equivalent(distinctiveness_range(valid_mat, dist_mat, 0.6),
-                    good_di_0.6)
+  expect_equal(distinctiveness_range(valid_mat, dist_mat, 0.6),
+               good_di_0.6)
 
   #  T ≥ 8/9
-  expect_equivalent(distinctiveness_range(valid_mat, dist_mat, 0.9),
-                    good_di_0.9)
+  expect_equal(distinctiveness_range(valid_mat, dist_mat, 0.9),
+               good_di_0.9)
 
-  expect_equivalent(distinctiveness_range(valid_mat, dist_mat, 4, TRUE),
-                    good_di_relative_max)
+  expect_equal(distinctiveness_range(valid_mat, dist_mat, 4, TRUE),
+               good_di_relative_max)
 })
 
 test_that("Correct Di computation with abundance",{
@@ -175,20 +167,20 @@ test_that("Correct Di computation with abundance",{
 
   # Check computation equal
   # When T < 1/9
-  expect_equivalent(distinctiveness_range(abund_mat, dist_mat, 0.1),
-                    good_di_0.1_ab)
+  expect_equal(distinctiveness_range(abund_mat, dist_mat, 0.1),
+               good_di_0.1_ab)
 
   # 1/9 ≤ T < 4/9
-  expect_equivalent(distinctiveness_range(abund_mat, dist_mat, 0.4),
-                    good_di_0.4_ab)
+  expect_equal(distinctiveness_range(abund_mat, dist_mat, 0.4),
+               good_di_0.4_ab)
 
   # 4/9 ≤ T < 8/9
-  expect_equivalent(distinctiveness_range(abund_mat, dist_mat, 0.6),
-                    good_di_0.6_ab)
+  expect_equal(distinctiveness_range(abund_mat, dist_mat, 0.6),
+               good_di_0.6_ab)
 
   #  T ≥ 8/9
-  expect_equivalent(distinctiveness_range(abund_mat, dist_mat, 0.9),
-                    good_di_0.9_ab)
+  expect_equal(distinctiveness_range(abund_mat, dist_mat, 0.9),
+               good_di_0.9_ab)
 })
 
 test_that("Distinctiveness works with sparse matrices", {
@@ -196,37 +188,37 @@ test_that("Distinctiveness works with sparse matrices", {
 
   ## Presence-absence
   # When T < 1/9
-  expect_equivalent(distinctiveness_range(sparse_mat, dist_mat, 0.1),
-                    as(good_di_0.1, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_mat, dist_mat, 0.1),
+               as(good_di_0.1, "dgeMatrix"))
 
   # 1/9 ≤ T < 4/9
-  expect_equivalent(distinctiveness_range(sparse_mat, dist_mat, 0.4),
-                    as(good_di_0.4, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_mat, dist_mat, 0.4),
+               as(good_di_0.4, "dgeMatrix"))
 
   # 4/9 ≤ T < 8/9
-  expect_equivalent(distinctiveness_range(sparse_mat, dist_mat, 0.6),
-                    as(good_di_0.6, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_mat, dist_mat, 0.6),
+               as(good_di_0.6, "dgeMatrix"))
 
   #  T ≥ 8/9
-  expect_equivalent(distinctiveness_range(sparse_mat, dist_mat, 0.9),
-                    as(good_di_0.9, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_mat, dist_mat, 0.9),
+               as(good_di_0.9, "dgeMatrix"))
 
   ## Abundances
   # When T < 1/9
-  expect_equivalent(distinctiveness_range(sparse_ab_mat, dist_mat, 0.1),
-                    as(good_di_0.1_ab, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_ab_mat, dist_mat, 0.1),
+               as(good_di_0.1_ab, "dgeMatrix"))
 
   # 1/9 ≤ T < 4/9
-  expect_equivalent(distinctiveness_range(sparse_ab_mat, dist_mat, 0.4),
-                    as(good_di_0.4_ab, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_ab_mat, dist_mat, 0.4),
+               as(good_di_0.4_ab, "dgeMatrix"))
 
   # 4/9 ≤ T < 8/9
-  expect_equivalent(distinctiveness_range(sparse_ab_mat, dist_mat, 0.6),
-                    as(good_di_0.6_ab, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_ab_mat, dist_mat, 0.6),
+               as(good_di_0.6_ab, "dgeMatrix"))
 
   #  T ≥ 8/9
-  expect_equivalent(distinctiveness_range(sparse_ab_mat, dist_mat, 0.9),
-                    as(good_di_0.9_ab, "dgeMatrix"))
+  expect_equal(distinctiveness_range(sparse_ab_mat, dist_mat, 0.9),
+               as(good_di_0.9_ab, "dgeMatrix"))
 })