+++ NEWS.md
@@ -1,3 +1,23 @@
+# funrar 1.5.0
+
+## NEW FEATURES
+
+* Added a new vignette named "other_functions" to document recently added
+functions. You can access it through
+`vignette("other_functions", package = "funrar")`
+
+## INTERNAL CHANGES
+
+* Fix uses of the `Matrix` package for their next version
+* Reorganized documentation to please `goodpractice`
+
+## USER-FACING CHANGES
+
+* Remove warnings regarding relative abundances for the `distinctiveness_*()`
+functions as they can perfectly work with absolute abundances giving the same
+results.
+* Fix typo in name of sparse matrices vignette.
+
 # funrar 1.4.1
 
 ## BUG FIX